script: Move textcontrol and related structs into dom/textcontrol
Part of #38901

Testing: it compiles
