+++ components/script/dom/document/document_embedder_controls.rs
@@ -30,10 +30,10 @@ use webrender_api::units::{DeviceIntRect, DevicePoint};
 use crate::dom::activation::Activatable;
 use crate::dom::bindings::refcounted::Trusted;
 use crate::dom::bindings::trace::NoTrace;
+use crate::dom::html::form_controls::text_control::TextControlElement;
 use crate::dom::inputevent::HitTestResult;
 use crate::dom::iterators::ShadowIncluding;
 use crate::dom::node::{Node, NodeTraits};
-use crate::dom::textcontrol::TextControlElement;
 use crate::dom::types::{
     Element, HTMLAnchorElement, HTMLElement, HTMLImageElement, HTMLInputElement, HTMLSelectElement,
     HTMLTextAreaElement, Window,
